chatmem0-extension/backend/pyproject.toml
[project]
name = "chatmem0-backend"
version = "0.1.0"
description = "ChatMem0 Backend API"
requires-python = ">=3.8"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn==0.24.0",
    "python-multipart==0.0.6",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "sqlalchemy==2.0.23",
    "pydantic==2.5.0",
    "python-dotenv==1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "black",
    "isort",
    "flake8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
dev-dependencies = [
    "pytest",
    "pytest-asyncio",
    "black",
    "isort",
    "flake8",
]
